requirements.txt
numpy==1.14.6
pillow==5.3.0
tqdm==4.26.0
tensorboardX==1.4
torchsummary==1.5.1
imageio==2.1.2